+++ app/src/main/java/com/sopt/clody/presentation/ui/writediary/component/bottomsheet/DeleteWriteDiaryBottomSheet.kt
@@ -1,4 +1,4 @@
-package com.sopt.clody.presentation.ui.writediary.component
+package com.sopt.clody.presentation.ui.writediary.component.bottomsheet
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
